SAAS-5435 refactoring
