+++ schema/1.0/documentation-links.js
@@ -1,0 +1,21 @@
+'use strict';
+
+const docBaseUrl = process.env.DOCS_BASE_URL || 'https://codefresh.io/docs/docs/codefresh-yaml/steps';
+
+const DocumentationLinks = {
+    'freestyle': `${docBaseUrl}/freestyle/`,
+    'build': `${docBaseUrl}/build/`,
+    'push': `${docBaseUrl}/push/`,
+    'deploy': `${docBaseUrl}/deploy/`,
+    'git-clone': `${docBaseUrl}/git-clone/`,
+    'launch-composition': `${docBaseUrl}/launch-composition/`,
+    'pending-approval': `${docBaseUrl}/approval/`,
+};
+
+const IntegrationLinks = {
+    'git-clone': `https://codefresh.io/docs/docs/integrations/git-providers/`,
+    'push': `https://codefresh.io/docs/docs/deploy-to-kubernetes/add-kubernetes-cluster/`,
+    'deploy': `https://codefresh.io/docs/docs/docker-registries/external-docker-registries/`,
+};
+
+module.exports = { docBaseUrl, DocumentationLinks, IntegrationLinks };
